pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gc-tracker-fr"
version = "2025.04.13"
description = "A tool designed to monitor the availability and prices online."
authors = [
    { name = "Kevin FISCHER", email = "mr.kevin.fischer@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml",
    "gspread",
    "oauth2client",
    "requests",
    "beautifulsoup4",
    "playwright",
    "playwright-stealth",
    "setuptools",
    "colorama",
    "streamlit",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "black",
]
